third_party/blink/web_tests/http/tests/security/contentSecurityPolicy/sandbox-report-only-expected.txt

CONSOLE ERROR: The Content Security Policy directive 'sandbox' is ignored when delivered in a report-only policy.
CONSOLE MESSAGE: Script executed in iframe.
ALERT: PASS: Iframe was not in a unique origin
